+++ google2/CHANGELOG.md
@@ -12,9 +12,12 @@ Changed:
 - added mocks for `GKEService` and `KubernetesService`
 - optimized implementation of `GoogleStorageInterpreter.getBlobBody` to fully use streams
 - log only result row count for BigQuery queries
+- Expose `GoogleComputeService.fromCredential`
+- Added max retries to `SubscriberConfig`
 
 Added:
 - Add `detachDisk`
+- Add `streamUploadBlob`
 - Add `listPodStatus` to `KubernetesService`, returns statuses of all pods belonging to a k8s cluster
 
 SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-TRAVIS-REPLACE-ME"`