+++ util2/CHANGELOG.md
@@ -6,6 +6,6 @@ This file documents changes to the `workbench-util2` library, including notes on
 
 - Add `ConsoleLogger`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util2" % "0.1-df50246"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util2" % "0.1-92fcd96"`
 
 Moved a few utilities that depends on `circe`, `fs2` from `util` to `util2`